Implements optional result queue support
