+++ Source/RetryNetworkService.swift
@@ -61,11 +61,11 @@ public final class RetryNetworkService: NetworkServiceProviding {
     }
     
     @discardableResult
-    public func request<T: ResourceModeling>(_ resource: T, onCompletion: @escaping (T.Model) -> Void,
+    public func request<T: ResourceModeling>(queue: DispatchQueue, resource: T, onCompletion: @escaping (T.Model) -> Void,
                         onError: @escaping (DBNetworkStackError) -> Void) -> NetworkTaskRepresenting {
         let retryTask = RetryNetworkTask(maxmimumNumberOfRetries: numberOfRetries, idleTimeInterval: idleTimeInterval, shouldRetry: shouldRetry,
                                   onSuccess: onCompletion, onError: onError, retryAction: { completion, error in
-                                    return self.networkService.request(resource, onCompletion: completion, onError: error)
+                                    return self.networkService.request(queue: queue, resource, onCompletion: completion, onError: error)
         }, dispatchRetry: { [weak self] disptachTime, block in
             self?.dispatchRetry(disptachTime, block)
         })
